Extract db maintenance into db-maintenance.js
The previous commit pushed server/lib/db.js to 615 lines, over the
600-line hard cap enforced by check:file-size --strict. Move the
incremental aggregate UPSERT + retention scheduler into a sibling
module, mirroring the existing db-schema.js split. db.js stays focused
on the query API.
